theme_ti

theme_ti

Usage

theme_ti(
  title_font = "Roboto Condensed Bold",
  base_font = "Roboto Condensed",
  void = FALSE,
  show_grid_lines = TRUE,
  base_text_size = 11,
  background_color = TRUE,
  .colors = ti_colors
)

Arguments

title_font

The default title is "Roboto". To swich to the matching Serif font picked out in the design phase (Lora), use title_font = "fancy". To change to a different font (e.g. to match up with an external publication requirement), simply change as desired (e.g. \"Times New Roman\"). Make sure you've installed the fonts you want to use on your device first!

base_font

The default font is "Roboto"). If you want to use a different font (e.g. to match up with an external publication requirement), simply change as desired (e.g. "Arial"). Make sure you've installed the fonts you want to use on your device first!

void

Logical (TRUE/FALSE). If TRUE, all grid lines and axes are removed. This is useful when creating pie/donut charts.

show_grid_lines

Logical (TRUE/FALSE). If FALSE, all grid lines are removed but the axis text is retained.

base_text_size

Base text size in pt. The relative size of the title / subtitle text and of the margins is derived from this.

background_color

Logical. The default (TRUE) adds a light color to the background in line with the Teal Insights color palette. Change to FALSE for a white background.

.colors

Used to read in ti_colors. Please leave as is!

Value

No object is returned. Instead, the theme is applied to a ggplot plot.